Correcciones en el entity mapper y configuraciones

+++ Entities/Promocion.cs
@@ -8,7 +8,7 @@ namespace Entities
     {
         public int id { get; set; }
         public string nombre { get; set; }
-        public decimal porcentaje { get; set; }
+        public double porcentaje { get; set; }
         public string codigo { get; set; }
         public int cantidad { get; set; }
         public int id_comercio { get; set; }
